+++ internal/util/relations_test.go
@@ -26,15 +26,15 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on consumer only",
 			args: args{
 				relations: ForeignRelations{
-					Consumer: []string{"foo", "bar"},
+					Consumer: []FR{{Identifier: "foo"}, {Identifier: "bar"}},
 				},
 			},
 			want: []Rel{
 				{
-					Consumer: "foo",
+					Consumer: FR{Identifier: "foo"},
 				},
 				{
-					Consumer: "bar",
+					Consumer: FR{Identifier: "bar"},
 				},
 			},
 		},
@@ -42,15 +42,15 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on consumer group only",
 			args: args{
 				relations: ForeignRelations{
-					ConsumerGroup: []string{"foo", "bar"},
+					ConsumerGroup: []FR{{Identifier: "foo"}, {Identifier: "bar"}},
 				},
 			},
 			want: []Rel{
 				{
-					ConsumerGroup: "foo",
+					ConsumerGroup: FR{Identifier: "foo"},
 				},
 				{
-					ConsumerGroup: "bar",
+					ConsumerGroup: FR{Identifier: "bar"},
 				},
 			},
 		},
@@ -58,15 +58,15 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on service only",
 			args: args{
 				relations: ForeignRelations{
-					Service: []string{"foo", "bar"},
+					Service: []FR{{Identifier: "foo"}, {Identifier: "bar"}},
 				},
 			},
 			want: []Rel{
 				{
-					Service: "foo",
+					Service: FR{Identifier: "foo"},
 				},
 				{
-					Service: "bar",
+					Service: FR{Identifier: "bar"},
 				},
 			},
 		},
@@ -74,15 +74,15 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on routes only",
 			args: args{
 				relations: ForeignRelations{
-					Route: []string{"foo", "bar"},
+					Route: []FR{{Identifier: "foo"}, {Identifier: "bar"}},
 				},
 			},
 			want: []Rel{
 				{
-					Route: "foo",
+					Route: FR{Identifier: "foo"},
 				},
 				{
-					Route: "bar",
+					Route: FR{Identifier: "bar"},
 				},
 			},
 		},
@@ -90,22 +90,22 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on service and routes only",
 			args: args{
 				relations: ForeignRelations{
-					Route:   []string{"foo", "bar"},
-					Service: []string{"foo", "bar"},
+					Route:   []FR{{Identifier: "foo"}, {Identifier: "bar"}},
+					Service: []FR{{Identifier: "foo"}, {Identifier: "bar"}},
 				},
 			},
 			want: []Rel{
 				{
-					Service: "foo",
+					Service: FR{Identifier: "foo"},
 				},
 				{
-					Service: "bar",
+					Service: FR{Identifier: "bar"},
 				},
 				{
-					Route: "foo",
+					Route: FR{Identifier: "foo"},
 				},
 				{
-					Route: "bar",
+					Route: FR{Identifier: "bar"},
 				},
 			},
 		},
@@ -113,26 +113,26 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on combination of route and consumer",
 			args: args{
 				relations: ForeignRelations{
-					Route:    []string{"foo", "bar"},
-					Consumer: []string{"foo", "bar"},
+					Route:    []FR{{Identifier: "foo"}, {Identifier: "bar"}},
+					Consumer: []FR{{Identifier: "foo"}, {Identifier: "bar"}},
 				},
 			},
 			want: []Rel{
 				{
-					Consumer: "foo",
-					Route:    "foo",
+					Consumer: FR{Identifier: "foo"},
+					Route:    FR{Identifier: "foo"},
 				},
 				{
-					Consumer: "foo",
-					Route:    "bar",
+					Consumer: FR{Identifier: "foo"},
+					Route:    FR{Identifier: "bar"},
 				},
 				{
-					Consumer: "bar",
-					Route:    "foo",
+					Consumer: FR{Identifier: "bar"},
+					Route:    FR{Identifier: "foo"},
 				},
 				{
-					Consumer: "bar",
-					Route:    "bar",
+					Consumer: FR{Identifier: "bar"},
+					Route:    FR{Identifier: "bar"},
 				},
 			},
 		},
@@ -140,26 +140,26 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on combination of service and consumer",
 			args: args{
 				relations: ForeignRelations{
-					Service:  []string{"foo", "bar"},
-					Consumer: []string{"foo", "bar"},
+					Service:  []FR{{Identifier: "foo"}, {Identifier: "bar"}},
+					Consumer: []FR{{Identifier: "foo"}, {Identifier: "bar"}},
 				},
 			},
 			want: []Rel{
 				{
-					Consumer: "foo",
-					Service:  "foo",
+					Consumer: FR{Identifier: "foo"},
+					Service:  FR{Identifier: "foo"},
 				},
 				{
-					Consumer: "foo",
-					Service:  "bar",
+					Consumer: FR{Identifier: "foo"},
+					Service:  FR{Identifier: "bar"},
 				},
 				{
-					Consumer: "bar",
-					Service:  "foo",
+					Consumer: FR{Identifier: "bar"},
+					Service:  FR{Identifier: "foo"},
 				},
 				{
-					Consumer: "bar",
-					Service:  "bar",
+					Consumer: FR{Identifier: "bar"},
+					Service:  FR{Identifier: "bar"},
 				},
 			},
 		},
@@ -167,43 +167,43 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on combination of service,route and consumer",
 			args: args{
 				relations: ForeignRelations{
-					Consumer: []string{"c1", "c2"},
-					Route:    []string{"r1", "r2"},
-					Service:  []string{"s1", "s2"},
+					Consumer: []FR{{Identifier: "c1"}, {Identifier: "c2"}},
+					Route:    []FR{{Identifier: "r1"}, {Identifier: "r2"}},
+					Service:  []FR{{Identifier: "s1"}, {Identifier: "s2"}},
 				},
 			},
 			want: []Rel{
 				{
-					Consumer: "c1",
-					Service:  "s1",
+					Consumer: FR{Identifier: "c1"},
+					Service:  FR{Identifier: "s1"},
 				},
 				{
-					Consumer: "c1",
-					Service:  "s2",
+					Consumer: FR{Identifier: "c1"},
+					Service:  FR{Identifier: "s2"},
 				},
 				{
-					Consumer: "c1",
-					Route:    "r1",
+					Consumer: FR{Identifier: "c1"},
+					Route:    FR{Identifier: "r1"},
 				},
 				{
-					Consumer: "c1",
-					Route:    "r2",
+					Consumer: FR{Identifier: "c1"},
+					Route:    FR{Identifier: "r2"},
 				},
 				{
-					Consumer: "c2",
-					Service:  "s1",
+					Consumer: FR{Identifier: "c2"},
+					Service:  FR{Identifier: "s1"},
 				},
 				{
-					Consumer: "c2",
-					Service:  "s2",
+					Consumer: FR{Identifier: "c2"},
+					Service:  FR{Identifier: "s2"},
 				},
 				{
-					Consumer: "c2",
-					Route:    "r1",
+					Consumer: FR{Identifier: "c2"},
+					Route:    FR{Identifier: "r1"},
 				},
 				{
-					Consumer: "c2",
-					Route:    "r2",
+					Consumer: FR{Identifier: "c2"},
+					Route:    FR{Identifier: "r2"},
 				},
 			},
 		},
@@ -211,43 +211,43 @@ func TestGetCombinations(t *testing.T) {
 			name: "plugins on combination of service,route and consumer group",
 			args: args{
 				relations: ForeignRelations{
-					Route:         []string{"r1", "r2"},
-					Service:       []string{"s1", "s2"},
-					ConsumerGroup: []string{"cg1", "cg2"},
+					Route:         []FR{{Identifier: "r1"}, {Identifier: "r2"}},
+					Service:       []FR{{Identifier: "s1"}, {Identifier: "s2"}},
+					ConsumerGroup: []FR{{Identifier: "cg1"}, {Identifier: "cg2"}},
 				},
 			},
 			want: []Rel{
 				{
-					ConsumerGroup: "cg1",
-					Service:       "s1",
+					ConsumerGroup: FR{Identifier: "cg1"},
+					Service:       FR{Identifier: "s1"},
 				},
 				{
-					ConsumerGroup: "cg1",
-					Service:       "s2",
+					ConsumerGroup: FR{Identifier: "cg1"},
+					Service:       FR{Identifier: "s2"},
 				},
 				{
-					ConsumerGroup: "cg1",
-					Route:         "r1",
+					ConsumerGroup: FR{Identifier: "cg1"},
+					Route:         FR{Identifier: "r1"},
 				},
 				{
-					ConsumerGroup: "cg1",
-					Route:         "r2",
+					ConsumerGroup: FR{Identifier: "cg1"},
+					Route:         FR{Identifier: "r2"},
 				},
 				{
-					ConsumerGroup: "cg2",
-					Service:       "s1",
+					ConsumerGroup: FR{Identifier: "cg2"},
+					Service:       FR{Identifier: "s1"},
 				},
 				{
-					ConsumerGroup: "cg2",
-					Service:       "s2",
+					ConsumerGroup: FR{Identifier: "cg2"},
+					Service:       FR{Identifier: "s2"},
 				},
 				{
-					ConsumerGroup: "cg2",
-					Route:         "r1",
+					ConsumerGroup: FR{Identifier: "cg2"},
+					Route:         FR{Identifier: "r1"},
 				},
 				{
-					ConsumerGroup: "cg2",
-					Route:         "r2",
+					ConsumerGroup: FR{Identifier: "cg2"},
+					Route:         FR{Identifier: "r2"},
 				},
 			},
 		},
@@ -263,9 +263,9 @@ func BenchmarkGetCombinations(b *testing.B) {
 	b.Run("consumer groups", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			relations := ForeignRelations{
-				Route:         []string{"r1", "r2"},
-				Service:       []string{"s1", "s2"},
-				ConsumerGroup: []string{"cg1", "cg2"},
+				Route:         []FR{{Identifier: "r1"}, {Identifier: "r2"}},
+				Service:       []FR{{Identifier: "s1"}, {Identifier: "s2"}},
+				ConsumerGroup: []FR{{Identifier: "cg1"}, {Identifier: "cg2"}},
 			}
 
 			rels := relations.GetCombinations()
@@ -275,9 +275,9 @@ func BenchmarkGetCombinations(b *testing.B) {
 	b.Run("consumers", func(b *testing.B) {
 		for i := 0; i < b.N; i++ {
 			relations := ForeignRelations{
-				Route:    []string{"r1", "r2"},
-				Service:  []string{"s1", "s2"},
-				Consumer: []string{"c1", "c2", "c3"},
+				Route:    []FR{{Identifier: "r1"}, {Identifier: "r2"}},
+				Service:  []FR{{Identifier: "s1"}, {Identifier: "s2"}},
+				Consumer: []FR{{Identifier: "c1"}, {Identifier: "c2"}, {Identifier: "c3"}},
 			}
 
 			rels := relations.GetCombinations()